pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maya-legal-intelligence"
version = "0.1.0"
description = "Maya Hieroglyphic Legal System with Symbolic Intelligence Integration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "BENNY HARIANTO", email = "creatoropensource@gmail.com"}
]
maintainers = [
    {name = "BENNY HARIANTO", email = "creatoropensource@gmail.com"}
]
keywords = [
    "legal-tech",
    "maya-hieroglyphic",
    "symbolic-intelligence",
    "legal-analysis",
    "ai-legal",
    "multi-jurisdictional"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Legal Industry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Legal",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "transformers>=4.20.0",
    "torch>=1.12.0",
    "nltk>=3.7",
    "spacy>=3.4.0",
    "pydantic>=1.10.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.20.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "httpx>=0.24.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=1.0.0",
]
enterprise = [
    "redis>=4.5.0",
    "celery>=5.2.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.10.0",
    "psycopg2-binary>=2.9.0",
]

[project.urls]
Homepage = "https://github.com/CreatorOss/maya-legal-intelligence"
Documentation = "https://maya-legal-intelligence.readthedocs.io"
Repository = "https://github.com/CreatorOss/maya-legal-intelligence"
"Bug Tracker" = "https://github.com/CreatorOss/maya-legal-intelligence/issues"
Changelog = "https://github.com/CreatorOss/maya-legal-intelligence/blob/main/CHANGELOG.md"

[project.scripts]
maya-legal = "maya_legal_intelligence.cli:main"
maya-analyze = "maya_legal_intelligence.analyzer:cli_main"
maya-server = "maya_legal_intelligence.server:run_server"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
maya_legal_intelligence = ["data/*.json", "models/*.pkl", "templates/*.html"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["maya_legal_intelligence"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=maya_legal_intelligence",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src/maya_legal_intelligence"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]